make the Micron driver robust to lossy communication
Since this is a request/response device, the driver will
stop functioning if there are some errors on the communication
line. Add a "retry" timeout to ensure robust behaviour even
in presence of packet loss.

While an adequate choice of IO read timeout would allow to handle
such cases in a "periodic" device, a request/response device is
sensitive to a single packet loss. This makes sure that we get a
proper behaviour, that is allow for packet loss while still triggering
IO timeout if no data ever comes.